+++ packages/calendar/src/components/period-slider/utils.ts
@@ -20,7 +20,12 @@ import { formatDate, monthName } from '../../utils';
 
 import { PeriodType } from '.';
 
-export const formatPeriod = (valueFrom: Date, valueTo: Date, periodType: PeriodType) => {
+export const formatPeriod = (
+    valueFrom: Date,
+    valueTo: Date,
+    periodType: PeriodType,
+    showCurrentYear?: boolean,
+) => {
     if (periodType === 'day') {
         if (isToday(valueFrom)) return 'Сегодня';
         if (isYesterday(valueFrom)) return 'Вчера';
@@ -29,11 +34,14 @@ export const formatPeriod = (valueFrom: Date, valueTo: Date, periodType: PeriodT
     }
 
     if (periodType === 'month') {
+        const monthAndYear = `${monthName(valueFrom)} ${getYear(valueFrom)}`;
+
+        if (showCurrentYear) {
+            return monthAndYear;
+        }
         const year = getYear(valueFrom);
 
-        return year === getYear(new Date())
-            ? monthName(valueFrom)
-            : `${monthName(valueFrom)} ${getYear(valueFrom)}`;
+        return year === getYear(new Date()) ? monthName(valueFrom) : monthAndYear;
     }
 
     if (periodType === 'quarter') {